+++ frontend/src/components/Header/Header.styled.ts
@@ -1,0 +1,13 @@
+// TODO: 검색결과 화면으로 변경시 다른 스타일이 적용되어야 함
+import heroImage from "assets/hero-img.png";
+
+const indexHeaderStyle = {
+  maxwidth: "1440px",
+  height: "640px",
+  backgroundImage: `url(${heroImage})`,
+  backgroundRepeat: "no-repeat",
+  backgroundPosition: "center bottom",
+  backgroundSize: "cover",
+};
+
+export default indexHeaderStyle;
