+++ frontend/src/components/Header/Header.tsx
@@ -2,20 +2,9 @@ import { Grid } from "@mui/material";
 import Box from "@mui/material/Box";
 import Container from "@mui/material/Container";
 
-import heroImage from "assets/hero-img.png";
-
 import GNB from "./GNB/GNB";
+import indexHeaderStyle from "./Header.styled";
 import UserMenu from "./UserMenu/UserMenu";
-
-// TODO: 검색결과 화면으로 변경시 다른 스타일이 적용되어야 함
-const indexHeaderStyle = {
-  maxwidth: "1440px",
-  height: "640px",
-  backgroundImage: `url(${heroImage})`,
-  backgroundRepeat: "no-repeat",
-  backgroundPosition: "center bottom",
-  backgroundSize: "cover",
-};
 
 const LogoArea = () => {
   return (
@@ -30,8 +19,13 @@ const LogoArea = () => {
 const Header = () => {
   return (
     <Box component="header" sx={indexHeaderStyle}>
-      <Container maxWidth="xl" sx={{ height: "100px" }}>
-        <Grid container spacing={2} columns={12}>
+      <Container maxWidth="xl" sx={{ height: ({ size }) => size.navBarHeight }}>
+        <Grid
+          container
+          spacing={2}
+          columns={12}
+          sx={{ height: ({ size }) => size.fullSize }}
+        >
           <LogoArea />
           <GNB />
           <UserMenu />
